MAINT - Split TestCommon to SSOCommon, SLOCommon, and Constants
