+++ ctk/common/src/main/kotlin/org/codice/compliance/utils/XMLDecryptor.kt
@@ -19,9 +19,6 @@ import org.apache.xml.security.encryption.XMLCipher
 import org.bouncycastle.jce.provider.BouncyCastleProvider
 import org.codice.compliance.children
 import org.codice.compliance.recursiveChildren
-import org.codice.compliance.utils.TestCommon.Companion.KEYSTORE_PASSWORD
-import org.codice.compliance.utils.TestCommon.Companion.PRIVATE_KEY_ALIAS
-import org.codice.compliance.utils.TestCommon.Companion.PRIVATE_KEY_PASSWORD
 import org.codice.compliance.utils.TestCommon.Companion.getCurrentSPHostname
 import org.w3c.dom.Document
 import org.w3c.dom.Element
