+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/NameIDPolicyVerifier.kt
@@ -20,11 +20,11 @@ import org.codice.compliance.SAMLCore_3_4_1_1_b
 import org.codice.compliance.attributeText
 import org.codice.compliance.children
 import org.codice.compliance.recursiveChildren
-import org.codice.compliance.utils.TestCommon.Companion.ASSERTION
-import org.codice.compliance.utils.TestCommon.Companion.FORMAT
-import org.codice.compliance.utils.TestCommon.Companion.ENCRYPTED_ID
-import org.codice.compliance.utils.TestCommon.Companion.SP_NAME_QUALIFIER
-import org.codice.compliance.utils.TestCommon.Companion.SUBJECT
+import org.codice.compliance.utils.ASSERTION
+import org.codice.compliance.utils.ENCRYPTED_ID
+import org.codice.compliance.utils.FORMAT
+import org.codice.compliance.utils.SP_NAME_QUALIFIER
+import org.codice.compliance.utils.SUBJECT
 import org.opensaml.saml.saml2.core.NameIDPolicy
 import org.w3c.dom.Node
