+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/assertions/StatementVerifier.kt
@@ -20,9 +20,9 @@ import org.codice.compliance.SAMLCore_2_7_4_a
 import org.codice.compliance.attributeNode
 import org.codice.compliance.children
 import org.codice.compliance.recursiveChildren
-import org.codice.compliance.utils.TestCommon.Companion.ASSERTION
-import org.codice.compliance.utils.TestCommon.Companion.AUTHN_STATEMENT
-import org.codice.compliance.utils.TestCommon.Companion.SUBJECT
+import org.codice.compliance.utils.ASSERTION
+import org.codice.compliance.utils.AUTHN_STATEMENT
+import org.codice.compliance.utils.SUBJECT
 import org.codice.compliance.verification.core.CommonDataTypeVerifier
 import org.w3c.dom.Node
