+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/profile/SingleSignOnProfileVerifier.kt
@@ -29,15 +29,15 @@ import org.codice.compliance.SAMLProfiles_4_1_4_2_k
 import org.codice.compliance.attributeNode
 import org.codice.compliance.attributeText
 import org.codice.compliance.children
-import org.codice.compliance.utils.TestCommon.Companion.ASSERTION
-import org.codice.compliance.utils.TestCommon.Companion.AUDIENCE
-import org.codice.compliance.utils.TestCommon.Companion.AUTHN_STATEMENT
-import org.codice.compliance.utils.TestCommon.Companion.BEARER
-import org.codice.compliance.utils.TestCommon.Companion.ENTITY
-import org.codice.compliance.utils.TestCommon.Companion.FORMAT
+import org.codice.compliance.utils.ASSERTION
+import org.codice.compliance.utils.AUDIENCE
+import org.codice.compliance.utils.AUTHN_STATEMENT
+import org.codice.compliance.utils.BEARER
+import org.codice.compliance.utils.ENTITY
+import org.codice.compliance.utils.FORMAT
+import org.codice.compliance.utils.SUBJECT
+import org.codice.compliance.utils.SUBJECT_CONFIRMATION
 import org.codice.compliance.utils.TestCommon.Companion.REQUEST_ID
-import org.codice.compliance.utils.TestCommon.Companion.SUBJECT
-import org.codice.compliance.utils.TestCommon.Companion.SUBJECT_CONFIRMATION
 import org.codice.compliance.utils.TestCommon.Companion.currentSPIssuer
 import org.codice.compliance.utils.TestCommon.Companion.getServiceUrl
 import org.codice.compliance.utils.TestCommon.Companion.idpMetadata
