+++ ctk/common/src/test/kotlin/org/codice/compilance/verification/core/SamlDefinedIdentifiersVerifierSpec.kt
@@ -26,12 +26,12 @@ import org.codice.compliance.SAMLCore_8_3_6_b
 import org.codice.compliance.SAMLCore_8_3_7_a
 import org.codice.compliance.SAMLCore_8_3_7_d
 import org.codice.compliance.SAMLCore_8_3_8_a
-import org.codice.compliance.utils.TestCommon.Companion.ASSERTION_NAMESPACE
-import org.codice.compliance.utils.TestCommon.Companion.ENTITY
-import org.codice.compliance.utils.TestCommon.Companion.PERSISTENT_ID
-import org.codice.compliance.utils.TestCommon.Companion.PROTOCOL_NAMESPACE
-import org.codice.compliance.utils.TestCommon.Companion.TRANSIENT_ID
-import org.codice.compliance.utils.TestCommon.Companion.VERSION
+import org.codice.compliance.utils.ASSERTION_NAMESPACE
+import org.codice.compliance.utils.ENTITY
+import org.codice.compliance.utils.PERSISTENT_ID
+import org.codice.compliance.utils.PROTOCOL_NAMESPACE
+import org.codice.compliance.utils.TRANSIENT_ID
+import org.codice.compliance.utils.VERSION
 import org.codice.compliance.verification.core.SamlDefinedIdentifiersVerifier
 import org.codice.compliance.verification.core.SamlDefinedIdentifiersVerifier.Companion.ATTRIBUTE_NAME_FORMAT_BASIC
 import org.codice.compliance.verification.core.SamlDefinedIdentifiersVerifier.Companion.ATTRIBUTE_NAME_FORMAT_UNSPECIFIED
