+++ ctk/idp/src/main/kotlin/org/codice/compliance/web/slo/RedirectSLOTest.kt
@@ -17,14 +17,14 @@ import com.jayway.restassured.RestAssured
 import io.kotlintest.specs.StringSpec
 import org.apache.cxf.rs.security.saml.sso.SSOConstants.SAML_REQUEST
 import org.apache.cxf.rs.security.saml.sso.SSOConstants.SAML_RESPONSE
-import org.codice.compliance.utils.TestCommon.Companion.EXAMPLE_RELAY_STATE
-import org.codice.compliance.utils.TestCommon.Companion.PARTIAL_LOGOUT
-import org.codice.compliance.utils.TestCommon.Companion.createDefaultLogoutRequest
-import org.codice.compliance.utils.TestCommon.Companion.createDefaultLogoutResponse
+import org.codice.compliance.utils.EXAMPLE_RELAY_STATE
+import org.codice.compliance.utils.PARTIAL_LOGOUT
+import org.codice.compliance.utils.SLOCommon.Companion.createDefaultLogoutRequest
+import org.codice.compliance.utils.SLOCommon.Companion.createDefaultLogoutResponse
+import org.codice.compliance.utils.SLOCommon.Companion.loginAndGetCookies
+import org.codice.compliance.utils.SLOCommon.Companion.sendRedirectLogoutMessage
 import org.codice.compliance.utils.TestCommon.Companion.encodeRedirectRequest
-import org.codice.compliance.utils.TestCommon.Companion.loginAndGetCookies
 import org.codice.compliance.utils.TestCommon.Companion.logoutRequestRelayState
-import org.codice.compliance.utils.TestCommon.Companion.sendRedirectLogoutMessage
 import org.codice.compliance.utils.TestCommon.Companion.useDSAServiceProvider
 import org.codice.compliance.utils.TestCommon.Companion.useDefaultServiceProvider
 import org.codice.compliance.utils.determineBinding
